API/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-api-python-client
google.generativeai
fastapi


